Refactoring main functionality in messageHandler struct
